src/requirements.txt
cycler==0.10.0
docopt==0.6.2
imageio==2.9.0
joblib==1.0.1
kiwisolver==1.3.2
matplotlib==3.4.3
networkx==2.6.2
nibabel==3.2.1
numpy==1.21.2
packaging==21.0
Pillow==8.4.0
pykwalify==1.8.0
pyparsing==2.4.7
pyradiomics==3.0.1
python-dateutil==2.8.2
PyWavelets==1.1.1
ruamel.yaml==0.17.14
ruamel.yaml.clib==0.2.6
scikit-image==0.18.3
scikit-learn==0.24.2
scipy==1.7.1
SimpleITK==2.1.0
six==1.16.0
sklearn==0.0
threadpoolctl==2.2.0
tifffile==2021.8.8
